myprojects/_golden_copy/pipeline-snowpark/source_code/pyproject.toml
[project]
name = "pipeline-dimcustomer-snowpark"
version = "0.1.0"
description = "Snowpark Python data pipeline - converted from PySpark"
requires-python = ">=3.10,<3.12"
dependencies = [
    "snowflake-snowpark-python>=1.20.0",
    "pandas>=2.0.0",
]
